Add files via upload
The process previously called Deconvolution is now called ConvTranspose, and the old name is now considered incorrect, so the names of the related layers have been changed. However, since many older programs are affected, the old names will be kept for compatibility reasons.
And also, the issue of the setting of gradients for the case padding is applied, backward method of ConvLayers are updated.